+++ crates/containerdebug/src/system_information/network.rs
@@ -1,11 +1,25 @@
-use hickory_resolver::{Resolver, system_conf::read_system_conf};
+use hickory_resolver::{
+    TokioResolver, name_server::TokioConnectionProvider, system_conf::read_system_conf,
+};
 use local_ip_address::list_afinet_netifas;
 use serde::Serialize;
 use std::{
     collections::{BTreeSet, HashMap},
     net::IpAddr,
+    sync::LazyLock,
     time::Duration,
 };
+use tokio::task::JoinSet;
+
+static GLOBAL_DNS_RESOLVER: LazyLock<TokioResolver> = LazyLock::new(|| {
+    let (resolver_config, mut resolver_opts) =
+        read_system_conf().expect("failed to read system resolv config");
+    resolver_opts.timeout = Duration::from_secs(5);
+
+    TokioResolver::builder_with_config(resolver_config, TokioConnectionProvider::default())
+        .with_options(resolver_opts)
+        .build()
+});
 
 /// Captures all system network information, including network interfaces,
 /// and the results of reverse and forward DNS lookups.
@@ -18,16 +32,7 @@ pub struct SystemNetworkInfo {
 
 impl SystemNetworkInfo {
     #[tracing::instrument(name = "SystemNetworkInfo::collect")]
-    pub fn collect() -> SystemNetworkInfo {
-        /*
-        let resolver = Resolver::from_system_conf()
-            .map_err(|e| e.to_string())
-            .unwrap();
-         */
-        let (resolver_config, mut resolver_opts) = read_system_conf().unwrap();
-        resolver_opts.timeout = Duration::from_secs(5);
-        let resolver = Resolver::new(resolver_config, resolver_opts).unwrap();
-
+    pub async fn collect() -> SystemNetworkInfo {
         let interfaces = match list_afinet_netifas() {
             Ok(netifs) => {
                 let mut interface_map = std::collections::HashMap::new();
@@ -56,12 +61,19 @@ impl SystemNetworkInfo {
             }
         };
 
-        let ip_set: BTreeSet<IpAddr> = interfaces.values().flatten().copied().collect();
-        tracing::info!(network.addresses.ip = ?ip_set, "ip addresses");
+        let ips: BTreeSet<IpAddr> = interfaces.values().flatten().copied().collect();
+        tracing::info!(network.addresses.ip = ?ips, "ip addresses");
 
-        let reverse_lookups: HashMap<IpAddr, Vec<String>> = ip_set
+        let mut reverse_lookups = JoinSet::new();
+        for ip in ips {
+            reverse_lookups
+                .spawn(async move { (ip, GLOBAL_DNS_RESOLVER.reverse_lookup(ip).await) });
+        }
+        let reverse_lookups: HashMap<IpAddr, Vec<String>> = reverse_lookups
+            .join_all()
+            .await
             .into_iter()
-            .filter_map(|ip| match resolver.reverse_lookup(ip) {
+            .filter_map(|(ip, reverse_lookup)| match reverse_lookup {
                 Ok(result) => {
                     let hostnames = result
                         .into_iter()
@@ -84,9 +96,20 @@ impl SystemNetworkInfo {
         let hostname_set: BTreeSet<String> = reverse_lookups.values().flatten().cloned().collect();
         tracing::info!(network.addresses.hostname = ?hostname_set, "hostnames");
 
-        let forward_lookups: HashMap<String, Vec<IpAddr>> = hostname_set
+        let mut forward_lookups = JoinSet::new();
+        for hostname in hostname_set {
+            forward_lookups.spawn(async move {
+                (
+                    hostname.clone(),
+                    GLOBAL_DNS_RESOLVER.lookup_ip(hostname).await,
+                )
+            });
+        }
+        let forward_lookups: HashMap<String, Vec<IpAddr>> = forward_lookups
+            .join_all()
+            .await
             .into_iter()
-            .filter_map(|hostname| match resolver.lookup_ip(hostname.clone()) {
+            .filter_map(|(hostname, forward_lookup)| match forward_lookup {
                 Ok(result) => {
                     let ips = result.iter().collect();
                     tracing::info!(hostname, ?ips, "performed forward DNS lookup for hostname");